refactor: 💡 clean up before changes
