+++ server/src/internal/api/rewards/handlers/referrals/handleDeleteRewardProgram.ts
@@ -1,0 +1,22 @@
+import { RewardProgramService } from "@/internal/rewards/RewardProgramService.js";
+import { routeHandler } from "@/utils/routerUtils.js";
+
+export default async (req: any, res: any) =>
+	routeHandler({
+		req,
+		res,
+		action: "delete reward scheme",
+		handler: async (req, res) => {
+			const { orgId, env, db } = req;
+			const { id } = req.params;
+
+			const rewardProgram = await RewardProgramService.delete({
+				db,
+				id,
+				orgId,
+				env,
+			});
+
+			return res.status(200).json(rewardProgram);
+		},
+	});
