+++ server/src/internal/api/rewards/handlers/referrals/index.ts
@@ -1,0 +1,3 @@
+export { default as handleCreateRewardProgram } from "./handleCreateRewardProgram.js";
+export { default as handleCreateTrigger } from "./handleCreateTrigger.js";
+export { default as handleDeleteRewardProgram } from "./handleDeleteRewardProgram.js";
