+++ server/src/internal/api/rewards/handlers/rewards/index.ts
@@ -1,0 +1,6 @@
+import handleCreateCoupon from "./handleCreateCoupon.js";
+import handleDeleteCoupon from "./handleDeleteCoupon.js";
+import handleGetCoupon from "./handleGetCoupon.js";
+import handleUpdateCoupon from "./handleUpdateCoupon.js";
+
+export { handleCreateCoupon, handleDeleteCoupon, handleGetCoupon, handleUpdateCoupon };
